requirements.txt
edge-tts>=6.1.9
pyyaml>=6.0
aiofiles>=23.0
